+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/CommonAbstracts/TextSource.kt
@@ -1,0 +1,6 @@
+package com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts
+
+interface TextSource {
+    val asMarkdownSource: String
+    val asHtmlSource: String
+}