+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/CodeTextMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.CodeTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.codeHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.codeMarkdown
 
@@ -7,7 +9,4 @@ data class CodeTextMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.codeMarkdown()
-    override val asHtmlSource: String = sourceString.codeHTML()
-}
+) : MessageEntity, TextSource by CodeTextSource(sourceString)