+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/EMailMessageEntity.kt
@@ -1,13 +1,12 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.EMailTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.emailHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.emailMarkdown
 
-class EMailMessageEntity(
+data class EMailMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.emailMarkdown()
-    override val asHtmlSource: String = sourceString.emailHTML()
-}
+) : MessageEntity, TextSource by EMailTextSource(sourceString)