+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/ItalicTextMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.ItalicTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.italicHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.italicMarkdown
 
@@ -7,7 +9,4 @@ data class ItalicTextMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.italicMarkdown()
-    override val asHtmlSource: String = sourceString.italicHTML()
-}
+) : MessageEntity, TextSource by ItalicTextSource(sourceString)