+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/PhoneNumberMessageEntity.kt
@@ -1,5 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.PhoneNumberTextSource
 import com.github.insanusmokrassar.TelegramBotAPI.utils.phoneHTML
 import com.github.insanusmokrassar.TelegramBotAPI.utils.phoneMarkdown
 
@@ -7,7 +9,4 @@ data class PhoneNumberMessageEntity(
     override val offset: Int,
     override val length: Int,
     override val sourceString: String
-) : MessageEntity {
-    override val asMarkdownSource: String = sourceString.phoneMarkdown()
-    override val asHtmlSource: String = sourceString.phoneHTML()
-}
+) : MessageEntity, TextSource by PhoneNumberTextSource(sourceString)